refactor: 모달 구조 분리 및 스타일/로직 리팩토링
- handleDescriptionModal()이 여러 번 불리는데 그 안에서 subtype이 여전히 description으로 남아있는 경우가 생겨서 문제가 발생해서 기존 'common' 모달 구조를 제거하고 data-type 기반으로 모달 구조 분리
  (설명, 설정, 점수판, 일시정지, 닉네임 등록 모달 각각 개별 dialog로 정의)
- modal.js 수정: 각 모달 타입에 따라 전용 핸들러 연결 및 showModal 처리
- description.js 등 개별 모달 핸들러 로직에서 data-subtype 검사 제거
- modal.css 리팩토링: 각 모달의 크기, 여백, 버튼 위치 등을 type에 따라 분리
- 공통 레이아웃(.modal, .modal-inner 등)은 그대로 유지하되 모듈화 강화
